Remove defined(__cplusplus) tests in C++ code.
This header is a C++ header (it contains keywords such as 'class'
and 'public'). It is not necessary to test defined(__cplusplus).
That test is appropriate in a C header that may be included by C++
code.

R=henrik.lundin@webrtc.org, jan.skoglund@webrtc.org, sprang@webrtc.org
BUG=none

Review URL: https://webrtc-codereview.appspot.com/38899004

Cr-Original-Commit-Position: refs/heads/master@{#8256}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: 74d27884aff095466cae76d2dcf90cb993321c14