Update the dummy file_audio_video_device to allow empty file name
Landing this on behalf of malmnas@.

The semantics is as follows:

* if the output filename is empty, then don't log to file
* if the input filename is empty, then don't stream any audio from file

This is useful for long running tests with multiple participants.
With logging turned on, having 10 bots running for 2 hours results in
more then 7 GB of data.

BUG=None
R=henrika@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/41219004

Cr-Original-Commit-Position: refs/heads/master@{#8691}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: 5c9f69f0afcc9bba998ba8ca6b40df15592a79c5