Testing that waiting for a condition variable waits.
Added a test that verifies that waiting for a condition variable
actually waits for a non-zero time.

This used to fail due to a TSAN / CLANG bug, but this failure
is supposed to have been fixed.

This was originally https://webrtc-codereview.appspot.com/2145004

BUG=2259

Review URL: https://codereview.webrtc.org/1416873002

Cr-Original-Commit-Position: refs/heads/master@{#10341}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: 3866c4fb460ebcf95857dbee6ec0e8664bd3b16a