Fix expectation bug in the RTPSender unit test.
The current expectation for InsertPacket(...) uses WillRepeatedly, which accepts if the function is called zero or more times. This CL changes this to either a fixed number of calls, or at least a positive number of calls.

Review URL: https://codereview.webrtc.org/1585783003

Cr-Original-Commit-Position: refs/heads/master@{#11256}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: 5d332ac8ffc1ff8efb35abd12b913ec495873494