Android EGL: Synchronize calls to eglCreateContext
Synchronize calls to EGL10/EGL14.eglCreateContext on EglBase.lock. The
reason is that a deadlock between the remote render thread in
eglSwapBuffers and MediaCodecVideoEncoder eglCreateContext was observed.

The function calls that are now synchronized on EglBase.lock are:
eglCreateContext, eglMakeCurrent, eglSwapBuffers, and
SurfaceTexture.updateTexImage.

BUG=webrtc:5702

Review-Url: https://codereview.webrtc.org/1937933002
Cr-Original-Commit-Position: refs/heads/master@{#12603}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: ddf165393fcb6ba9b7e7c0219852b5247d1937ed