Fixed a crash in Objective-C clients when data channel becomes closed.
Objective-C applications using the data channel could crash when the following conditions were met:
 - The application creates a data channel and sets a data channel delegate
 - The delegate is deallocated but the application never clears it
When this happens, the dealloc method in RTCDataChannel would reset the delegate to nil. The setDelegate: method immediately returns if the new delegate value is the same as the old one. Since the old one is a weak reference, it becomes nil, and the DataChannelDelegateAdapter is not unsubscribed from the native channel before it gets deleted.

The fix removes the two optimizations, and instead subscribes the adapter to the native data channel at creation time - and unsubscribes it at dealloc time. This makes it very easy to reason about the lifetime of the subscription. Removing the optimization should have little effect on performance, as applications typically set the delegate when the channel is created.

BUG=

Review-Url: https://codereview.webrtc.org/1957523006
Cr-Original-Commit-Position: refs/heads/master@{#12649}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: 630d9ba364db875ee5a4b78a6e7f1fbaa0719b10