Android: Make base interface for camera1 and camera2
This CL adds a new interface CameraVideoCapturer that extends VideoCapturer with a switchCamera() function. It also moves moves CameraEventsHandler, CameraStatistics, and CameraSwitchHandler from VideoCapturerAndroid to this new interface. The purpose is to prepare for a camera2 implementation that will use the same interfaces and helper class.

BUG=webrtc:5519

Review-Url: https://codereview.webrtc.org/1895483002
Cr-Original-Commit-Position: refs/heads/master@{#12723}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: 6bdacaddfb18edef1f0cdd778209f6b05a8f9210