Don't re-determine ICE role on an ICE restart.
This only causes an increased likelihood of role conflicts as each peer
is picking a new role. It also means that if ICE is restarted for only
one media stream, roles can be different across media streams (which isn't
even allowed).

For more explanation of why this is unnecessary and should be changed,
see this discussion:
https://mailarchive.ietf.org/arch/msg/ice/C0_QRCTNcwtvUF12y28jQicPR10

Review-Url: https://codereview.webrtc.org/2046123003
Cr-Original-Commit-Position: refs/heads/master@{#13077}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: 73fbcf9237b1109283ef590b24188b1d64144358