A missing path separator caused aecdump recordings not to be created in apprtc on Android.
The path separator was missing when the path for the aecdump
file was created. This CL adds that path separator.

Note that the change of the formatting of the rest of the
line was caused by "git cl format" (the clang automatic
formatting).

BUG=webrtc:5991

Review-Url: https://codereview.webrtc.org/2053263002
Cr-Original-Commit-Position: refs/heads/master@{#13121}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: 81d99b304969021cb1799ed6bfc19f2792959a53