Change GetPlayoutTimestamp message from Error to Info
This message was printed to the receiver's log on every single video
frame if the remote side is sending video but not audio. This could
happen if, for example, one participant creates RtpSenders for both
audio and video but doesn't set the audio track, or if the track is
set but RtpParameters.encodings[0].active == false.

This CL changes the trace level to StateInfo as it is the expected
behavior in this case.

R=deadbeef@webrtc.org, henrika@webrtc.org

Review URL: https://codereview.webrtc.org/2125183002 .

Cr-Original-Commit-Position: refs/heads/master@{#13408}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: 4c0536bd510a17c0d793ff3f811c25bee7f60fea