Return nil from RTCPeerConnectionFactory when creation fails
RTCPeerConnectionFactory.createPeerConnection did not check the return
value of the native createPeerConnection() - so when the native PC
fails to be created, it could end up attempting to use a null pointer.

The change makes it return nil when the creation fails. The application
can then detect and respond to the failure.

Review-Url: https://codereview.webrtc.org/2240633004
Cr-Original-Commit-Position: refs/heads/master@{#13732}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: 588783adcd2bcf1a788cfd318e1394f241e9a78f