Cleanup of the AudioDeviceBuffer class.
WebRTC works on 10ms buffer sizes in both directions but this class has contained
support for any size (with some limits) and for changes on the fly. It makes no sense to maintain such code and we have no tests to test it. This CL ensures that only 10ms audio buffers are supported and that nothing can be changed on the fly.

It also updates the style to follow the Google C++ style guide.

Finally, I remove very old (not tested and not maintained) support for file
handling since the code is never used. It was more or less dead code.

BUG=NONE
R=magjed@webrtc.org

Review URL: https://codereview.webrtc.org/2256833003 .

Cr-Original-Commit-Position: refs/heads/master@{#13833}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: cf327b45b9f5738950d4fca2b6a7b6030d508cdf