Fixed flaky VideoSendStreamTest::SupportsAbsoluteSendTime
This test failed on the memcheck bot:
https://build.chromium.org/p/client.webrtc/builders/Linux%20Memcheck/builds/6704/steps/video_engine_tests/logs/stdio

The test assumed that the absolute send time header extension can never
be zero. It's a timestamp truncated to 24 bits, and zero is not a
special value - so it can very rarely end up being precisely zero.

The fix makes the test wait for at least one packet having a non-zero send time.

I've considered changing the test to use a fake clock instead to ensure
that not only the value is non-zero, but that it indeed reflects the
system timestamp - but that involves changing a very large number of
files. Besides, other tests in this file don't verify values for header
extensions where zeroes are allowed.

NOTRY=true

Review-Url: https://codereview.webrtc.org/2307693002
Cr-Original-Commit-Position: refs/heads/master@{#14056}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: c3f3515f8ea78c0a8a5026639e605a9bdfda4add