Assume ProjectRootPath() equals ../.. in Desktop
This way we don't have to rely on the existence of DEPS, and the tests
can be run in swarming bots (which don't have a checkout and therefore
don't have a DEPS file).

This seems to be where Chromium is assumming the project root path to
be.

NOTRY=True
BUG=chromium:497757

Review-Url: https://codereview.webrtc.org/2340773002
Cr-Original-Commit-Position: refs/heads/master@{#14230}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: 8b28b8017f31183f24fa2815bd4d3d57b03a8060