iOS: Fix rotation bug for back camera
The rotation is currently incorrect for the back camera in landscape
mode. The reason is that device rotation needs to be reversed for the
back camera compared to the front camera. The camera sensor can also be
mounted with a specific orientation. So when front camera rotation goes
through: 0->90->180->270, back camera rotation goes in reverse:
180->90->0->270.

BUG=b/31984246,b/30651939

Review-Url: https://codereview.webrtc.org/2401033002
Cr-Original-Commit-Position: refs/heads/master@{#14580}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: 9e31cca8f49cddab997843ffb8a62df7af9cad78