Stop caching supported codecs in WebRtcVideoEngine2
We currently cache the result of GetSupportedCodecs in a member variable
|video_codecs_| in WebRtcVideoEngine2. This means we need to keep
|video_codecs_| and the result of GetSupportedCodecs in sync, which is
error prone. It's simpler to just call GetSupportedCodecs when we need
it, and we actually end up making fewer calls, so it's faster as well.
This CL also returns all std::vectors by-value instead of by-ref. Move
semantic together with in-place filtering of codecs actually end up with
fewer copies, and it's also simpler to not return references.

BUG=webrtc:6337

Review-Url: https://codereview.webrtc.org/2492473002
Cr-Original-Commit-Position: refs/heads/master@{#15007}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: 9f71ec5a3e3175751f4475b126cfda89767363f2