Fix BitrateControllerImpl not to ignore BW probe results mid-call.
Previously when BitrateControllerImpl::OnDelayBasedBweResult() is
called as result of a probe it was calling
bandwidth_estimation_.SetSendBitrate(), but not
UpdateDelayBasedEstimate(). As result SendSideBandwidthEstimation was
effectively ignoring probe results as it kept the old
delay_based_bitrate_bps_ value, which caps the resulting bitrate.

BUG=webrtc:6332,webrtc:6710

Review-Url: https://codereview.webrtc.org/2481383002
Cr-Original-Commit-Position: refs/heads/master@{#15071}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: 26b675625f150be1ebc1a2350d39e5116d214871