AGC: Add a histogram for clipping adjustment
This new histogram will log a value every time input clipping is
detected. The value is a boolean, with "true" meaning that the gain
was in fact adjusted in response to the detected clipping, and "false"
meaning that adjustment was not allowed due to kClippedLevelMin.

BUG=webrtc:6622

Review-Url: https://codereview.webrtc.org/2522543006
Cr-Original-Commit-Position: refs/heads/master@{#15196}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: 30a12fbbb9f89959b9a4f3f81b20a9dbb54b5c66