Fixed bug in ExtractFrameFromY4mFile API which was not extracting the frames correctly.
Issue: This API was calculating the file_header and frame_header offset only for the first frame which is not the right logic. We need to skip the file and frame header every time we extract a new frame.

Also added a unit test case which compares the extracted frame with the frame stored in text file.

BUG=webrtc:6761

NOPRESUBMIT=true
NOTRY=true

Review-Url: https://codereview.webrtc.org/2529923002
Cr-Original-Commit-Position: refs/heads/master@{#15260}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: b7636b4656d7f8c368963f2256dc2ef7b7ba89c8