Fixed bug in ExtractFrameFromY4mFile API which was not extracting the frames correctly.
Issue: This API was calculating the file_header and frame_header offset only for the first frame which is not the right logic. We need to skip the file and frame header every time we extract a new frame.

Also added a unit test case which compares the extracted frame with the frame stored in text file.

NOPRESUBMIT=true
NOTRY=true

BUG=webrtc:6761

Review-Url: https://codereview.webrtc.org/2532963002
Cr-Original-Commit-Position: refs/heads/master@{#15288}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: 1b5b22dc1797e38a0e6fcce59cf2b6ea92dfcccc