Make SurfaceTextureHelper and I420Frame public in Java.
This change makes the Java classes and constructors for
SurfaceTextureHelper and I420Frame public. This allows applications to
use the WebRTC CameraVideoCapturer to obtain raw frames, and to render
frames on the WebRTC VideoRenderer without having to pass them through a
VideoTrack - such as when using Quartc.

BUG=None.

Review-Url: https://codereview.webrtc.org/2544563002
Cr-Original-Commit-Position: refs/heads/master@{#15344}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: ed6e077263eefff2b7f8595c85b70ee3193bbe9f