Remove bitrate cap for AdaptiveVideoSource and increase other caps to 25 Mbps.
AdaptiveVideoSource is used in testing/simulations of the bandwidth estimator.

Nada's reaction to delay depends on the current bitrate and the configured max rate in a non-intuituve way. Increase the starting bitrate to compensate for the increased max bitrate. This is only used in unit tests.

BUG=webrtc:6807

# Presubmit warns about a lint error in bwe.h that's unrelated to my change. Fixing it is beyond the scope of this CL.
NOPRESUBMIT=true

Review-Url: https://codereview.webrtc.org/2542843003
Cr-Original-Commit-Position: refs/heads/master@{#15364}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: 182e4a4aff7125e2e9e3217b253bc4d1545f8bce