Only store sequence numbers for media stream in FlexFEC end-to-end test.
This should remove the test flakiness, as before this change there
could be collisions from sequence numbers coming from two sequence
number spaces (the media SSRC and the FlexFEC SSRC). The probability
of collisions was low, and hence the flakes were far between.

This change also reduces the packet loss to 5% (down from ~50%), in
order for the BWE to have an easier time to ramp up.

BUG=webrtc:6825
R=philipel@webrtc.org, mflodman@webrtc.org

Review-Url: https://codereview.webrtc.org/2554403003
Cr-Original-Commit-Position: refs/heads/master@{#15512}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: 3536463e7e808747ddc5da2dc0933dbec97ceeeb