Move all codec specific definitions from modules_include
This CL moves all codec specific definitions into their own
header files in the respective codec directory, and replaces
it with an include in the top level directory.

This is to facilitate getting the code out of the header files.

No behavioral changes are expected.

BUG=webrtc:6842

Review-Url: https://codereview.webrtc.org/2555993003
Cr-Original-Commit-Position: refs/heads/master@{#15623}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: 41286496cbf61c2ab66b2df243e00c206995de39