Relax parsing of a=bundle-only with a nonzero port.
This use of a=bundle-only is unspecified, but not disallowed, so it
should simply result in the attribute being ignored, not a parse
failure.

Note that older versions of Firefox may generate SDP with a=bundle-only
and a nonzero port, so this also fixes an interop issue. See:
https://bugzilla.mozilla.org/show_bug.cgi?id=1325991

BUG=webrtc:4674

Review-Url: https://codereview.webrtc.org/2609863003
Cr-Original-Commit-Position: refs/heads/master@{#15890}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: 12771a19103b8d8f037f67f41299a8256881764d