New simulators to test BWE at low bitrates (15-50kbps range).
I've tested both the old and the new delay estimators and they all work ok at 15 kbps. The new estimators are a bit slower to adapt down so the maximum delay is a bit higher, especially at lower bitrates.

None of the estimators work at 10 kbps, but that is likely because the bitrate controller is configured to never go below 10 kbps. This means that it's impossible to empty the queues after a capacity drop to 10kbps (regardless of what the estimators do), so the delay stays high until the capacity increases.

BUG=webrtc:7022

Review-Url: https://codereview.webrtc.org/2644463009
Cr-Original-Commit-Position: refs/heads/master@{#16233}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: 568c9e72d149662dab92a671bd3139e4ae7f167a