Fix neteq_speed_test.cc
After https://codereview.webrtc.org/2340773002,
the path from webrtc::test::ResourcePath in
webrtc/modules/audio_coding/neteq/tools/neteq_performance_test.cc is wrong.

It is
/path/to/repos/resources/audio_coding/testfile32kHz.pcm

It should be
/path/to/repos/webrtc-temp/src/resources/audio_coding/testfile32kHz.pcm.

The middle part is missing.

The reason this target is affected is because
webrtc::test::SetExecutablePath(argv[0]);
was not called.

That call is necessary for us to know that the test is being run from src/
and not from out/Default (as is assumed, when that function is not called.)

BUG=chromium:497757
R=kjellander@webrtc.org, henrik.lundin@webrtc.org

Review-Url: https://codereview.webrtc.org/2698743002
Cr-Original-Commit-Position: refs/heads/master@{#16641}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: 454c1d6a2305b7f0c8395b8e76b9a8aaaf8ee258