Further optimization of AudioVector::operator[]
This is a follow-up to https://codereview.webrtc.org/2670643007/. That
CL provided significant improvement to Mac, Linux and ARM-based
platforms, but failed to improve the performance for Windows. The
problem is that the MSVC compiler did not produce branch-free code for
that fix. This new change produces the same result for non-Windows
platforms, as well as introduces branch-free code for Windows.

H/t to kwiberg@ for providing the solution.

BUG=webrtc:7159

Review-Url: https://codereview.webrtc.org/2700633003
Cr-Original-Commit-Position: refs/heads/master@{#16649}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: 751589899b753d8ca56ac6e2ca2742596520cc8e