Fix rare race in ViEEncoderTest.UMACpuLimitedResolutionInPercent
The test aims to synchronize by waiting for the last sent frame, using
sink_.WaitForEncodedFrame(). Then it stops the vie_encoder instance and
finally reset the SendStatsProxy to trigger UMA stats to be registered.

The problem is that the callback that triggers WaitForEncodedFrame() to
release will call FrameSent() on the overuse detecter just after that,
and so it may try to update the stats proxy while it is being deleted.

To prevent this, simply destroy the vie_encoder instance too, so that
we know that the task queue has been stopped.

BUG=None

Review-Url: https://codereview.webrtc.org/2720183004
Cr-Original-Commit-Position: refs/heads/master@{#16914}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: f8ee65ead6e16cbb1c4311b9fd596a3435ec2f9a