Hiding problematic targets from chromium
Targets that depend on //third_party/gflags are problematic in
chromium because it does not use gflags and if a target depends on
any target in //webrtc/test/BUILD.gn then GN will try to parse the
BUILD.gn file. Gn is eager in this phase of the build files generation
and if a target is not buildable it will complain.

In chromium obviously all the targets that depend on
//third_party/gflags are not buildable.

To be able to enable 'gn check' on the webrtc/base and
webrtc/test directories we have to hide them from chromium.

BUG=webrtc:6828
NOTRY=True

Review-Url: https://codereview.webrtc.org/2845063002
Cr-Original-Commit-Position: refs/heads/master@{#17913}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: 9452c62f28d7159bfe87c5a06998ce86373610c1