Disable CpuTimeTest on instrumented builds.
The test relies on the ability to run a thread for a controllable
duration of time, which is unreliable on instrumented builds (TSan,
etc.).

Also adding a minimal test which can still be run on instrumented builds, so
they at least still exercise the code.

BUG=webrtc:7542
TBR=kjellander@webrtc.org

Review-Url: https://codereview.webrtc.org/2856473002
Cr-Original-Commit-Position: refs/heads/master@{#17961}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: 7311b246609973ae3980a747999e62e025e1e75f