This CL introduces a new APM sub-module named AGC2 that does not use the band split domain and only implements floating point operations (to avoid spectral leakage issues and unnecessary complexity).
The goal of this CL is adding the new sub-module into APM without providing an
implementation that could replace the existing gain control modules. The focus
is in fact on initialization, reset, and configuration of AGC2.

The module itself only applies a hard-coded gain value. This behavior will
change in the coming CLs.

BUG=webrtc:7494

Review-Url: https://codereview.webrtc.org/2848593002
Cr-Original-Commit-Position: refs/heads/master@{#18222}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: 3ec96df9070a0395c7d21d41c48bfe68b26cba92