Add media constraint-key constants for generating offers and answers
Similar to the existing constants for the media constraint-keys intended
for media sources, add the constants for the standard keys to generate
offers and answers.

This patch also adds a few comments to RTCMediaConstraints.h, to give
a better clue on the intended usage scope of declared media-constraint
keys and values.

BUG=webrtc:7722

Review-Url: https://codereview.webrtc.org/2908013002
Cr-Original-Commit-Position: refs/heads/master@{#18362}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: 7b3eca9940e4e9407df2024248a212063680b6d6