Remove webrtc deps from AppRTCMobile.
We want the example app to only link agains the framework. This ensures
that we are actually testing the framework, and that AppRTCMobile
doesn't require any other parts of WebRTC not included in the framework.

Bug: webrtc:7759
Change-Id: Ib04aae0bc3ab2a1a508eaf4a4f15c2d37f521598
Reviewed-on: https://chromium-review.googlesource.com/522722
Reviewed-by: Magnus Jedvert <magjed@webrtc.org>
Commit-Queue: Kári Tristan Helgason <kthelgason@webrtc.org>
Cr-Original-Commit-Position: refs/heads/master@{#18470}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: 946923a21fad8e7b98c27d53f13a32d9a8513749