What can't loopback test be more like full stack test?
It can; this CL makes it a lot closer, if not all the way to a merge.
Performance from video_loopback and screenshare_loopback should now
match what we're seeing in FullStackTest, which will make debugging and
assesment of quality differences much easier.

It also adds the ability to view all of the simulcast streams at once,
in separate windows.

BUG=webrtc:7694

Review-Url: https://codereview.webrtc.org/2946893003
Cr-Original-Commit-Position: refs/heads/master@{#18703}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: 1168fd4ed560176987116f06f5a3bd257d471d6e