Remove unneeded enum forward declaration
While building Chrome with the VC++ 2017 /permissive- flag I got a
warning about a forward declaration of enum RateControlRegion. Untyped
forward declarations of enums are illegal because the compiler doesn't
know what size to make them. The only reason this forward declaration is
legal is because it isn't needed (the type is already defined).

This was found because /permissive- (or, equivalently for this purpose,
/w14471) incorrectly fires on this forward declaration even though it is
legal.

BUG=chromium:736059

Review-Url: https://codereview.webrtc.org/2834753002
Cr-Original-Commit-Position: refs/heads/master@{#18741}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: 5e5f7e14b23ee93fb163d64229da5ec5708f7cf9